+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## v2.3.0
+
+#### Breaking changes
+- Changed the exit code from 1 to 0 for `ls` when used with an empty bucket. Exits with 1 if the bucket is non-existent. ([#722](https://github.com/peak/s5cmd/issues/722))
+
+#### Features
+- Added prefix and wildcard support to `cat` command. ([#716](https://github.com/peak/s5cmd/issues/716))
+
 ## v2.2.2 - 13 Sep 2023 
 
 #### Bugfixes